Commonize connections between Pinger and Grapher
